requirements.txt
groq>=0.13.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pgvector>=0.3.0
